parser: move stuff around, fix formatting
